Вынесла класс в отдельный файл
